Fix problem with old British times.
mktime(3) appears to try to do daylight savings corrections even with
isdst and gmtoff set to zero, causing problems for dates between 27
October 1968 and 31 October 1971 when the British were on GMT+1.  (There
may be other examples of problems, but this was the first found.)  Using
the non-standard but prevalent timegm(3) call fixes this, as it only
takes UTC as input and so doesn't need to do any such manipulations.
As a result, we no longer need to subtract the timezone offset.