Experimental rewrite of DateTime using astropy
DateTime.cc now calls an experimental Python implementation
(implementation is not included in this commit) that itself
calls Astropy.time.

All public C++ methods are now implemented as boost.python calls
to the corresponding Python methods. For this experiment the SWIG
interface is still built so that the previous tests can run through
the C++ system and back into Python.

The dateTime_1.cc test is removed because standalone C++ binaries
no longer work with DateTime as the current system does not attempt
to initialize the Python runtime.

In this commit some tests fail in dateTime.py.

Exception handling is broken in that exceptions from the Python
are converted to a generic pex_exception rather than appearing
as true Python exceptions to the Python caller.