Remove inPlace
It looks like we don't actually use the inPlace argument to the
PropertyList add and set methods anywhere in the stack besides the
daf_base unit tests. I think it was originally put in to accommodate the
addition of COMMENT and HISTORY items that might want to be placed at
the end of the PropertyList instead of grouped together, but the
implementation actually still groups all of these items together, just
at the end of the list instead of where they were.