Try to guess int vs long when type is not known
Python 3 does not distinguish long from int but PropertySet
and PropertyList can have different types. If something is added
to a container that already contains ints then assume Int, if
it already contains LongLong use LongLong. If the container
is empty choose based on the value.

We are careful not to include bool types in the guess work.

Adds some tests to ensure that boolean are handled properly
and that OverflowError is raised if the item is too large.