Add handmade api pages for cpp classes
Breathe's doxygennamespace directive dumped all classes/functions from a
namespace in-place into the reST file. This behaviour is inconsistent
with what astropy-helper's automodapi does, which is make a new page for
every class and function in a module. Having individual pages is a nice
usability feature since it's easy to get lost in what class is
currently being documented.

It seems that a  automodapi equivalent for breathe needs to be written.
In the meantime, I've simulated that directive by creating pages for
each class in lsst::daf::base.