Improved robustness of iso ctro test
The ISO string DateTime constructor demands a trailing "Z" for UTC
and forbids it for all other time systems. However, the test for that
was written backwards -- testing for non-UTC systems by name --
meaning it would need udpating if another time system was added.
I fixed that so the test truly is "if UTC then demand a trailing Z,
otherwise forbid one". Note: all time systems are checked later on,
so an invalid time system will get through the test in question
but be caught later.