Change PropertySet.getArray to return all types
In updating code to use `PropertySet.getArray` or `getScalar`,
instead of `get`, I realized that the best thing `getArray` can do
with items that are neither numeric or scalar (e.g. `PropertySet`
or `PropertyList`) is return them, preferably as a scalar.

This allows one to retrieve all known information about any
item in a container by calling `getArray`.
Formerly this was difficult: you would have to first determine
the data type (e.g. by calling `getScalar`) then call `getArray`
if the type was numeric or string.