Improve _propertyContainerGet for property cont.
The way _propertyContainerGet got a PropertySet or PropertyList
was by first trying to return as a PropertyList, then checking
the type and returning as a PropertySet if it matched. This was
inefficient because the type of a PropertyList is the same as a
PropertySet, so do that check first. Also, use isPropertySetPtr
for the test to make the code look a bit cleaner.