Enable hierarchical serialization of PropertySet
Previously a PropertySet was pickled/YAML-ified as a flat
dot-delimited set of keys, types and values. This resulted
in PropertyList losing its identity inside a PropertySet,
such that comments were stripped and on read the
List becomes a Set.  With this change only top level keys
are processed and pickle/yaml then processes any contained
PropertySet/List and the true hierarchy is recorded in the
files. This does not break backwards compatibility with
previous serializations.